networkAnalysisTool.py
This developed tool will help determining the network traffic by analyzing the transport layer protocols by Packet Sniffing method.
Select the required network interface of your network and the number of packets to sniff. It will show the traffic in bar graph and can be displayed as a csv file by giving the filepath location.